chore: moved <rh-menu-item> and <rh-menu-item-group> inside <rh-menu>
