+++ elements/rh-menu-dropdown/demo/drodown-item-with-icons.html
@@ -1,20 +1,23 @@
 <div id="menu-dropdown-container">
   <rh-menu-dropdown text="Basic toggle" info-icon-name="auto-light-dark-mode">
     <p slot='label'>Basic toggle</p>
-    <rh-menu-item> <rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Action</rh-menu-item>
-    <rh-menu-item href="#"> <rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Link</rh-menu-item>
-    <rh-menu-item disabled> <rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Disabled Action</rh-menu-item>
-    <rh-menu-item disabled href="#"><rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Disabled link</rh-menu-item>
-    <rh-menu-item disabled><rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Aria-disabled link</rh-menu-item>
-    <hr/>
-    <rh-menu-item><rh-icon slot='icon' set="ui" icon="document"></rh-icon> Separated action</rh-menu-item>
-    <rh-menu-item href="#"><rh-icon slot='icon' set="ui" icon="document"></rh-icon> Separated link</rh-menu-item>
+    <rh-menu>
+      <rh-menu-item> <rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Action</rh-menu-item>
+      <rh-menu-item href="#"> <rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Link</rh-menu-item>
+      <rh-menu-item disabled> <rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Disabled Action</rh-menu-item>
+      <rh-menu-item disabled href="#"><rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Disabled link</rh-menu-item>
+      <rh-menu-item disabled><rh-icon slot='icon' set="ui" icon="profile"></rh-icon> Aria-disabled link</rh-menu-item>
+      <hr/>
+      <rh-menu-item><rh-icon slot='icon' set="ui" icon="document"></rh-icon> Separated action</rh-menu-item>
+      <rh-menu-item href="#"><rh-icon slot='icon' set="ui" icon="document"></rh-icon> Separated link</rh-menu-item>
+    </rh-menu>
   </rh-menu-dropdown>
 </div>
 
 <script type="module">
   import '@rhds/elements/rh-menu-dropdown/rh-menu-dropdown.js';
-  import '@rhds/elements/rh-menu-dropdown/rh-menu-item.js';
+  import '@rhds/elements/rh-menu/rh-menu-item.js';
+  import '@rhds/elements/rh-menu/rh-menu.js';
 
   const dropdown = document.querySelector('rh-menu-dropdown');
   dropdown.addEventListener('select', e => {
