+++ elements/rh-menu-dropdown/demo/info-action-icon-only.html
@@ -1,20 +1,23 @@
 <div id="menu-dropdown-container">
   <rh-menu-dropdown>
     <rh-icon slot='label' set="ui" icon="auto-light-dark-mode"></rh-icon>
-    <rh-menu-item>Action</rh-menu-item>
-    <rh-menu-item href="#">Link</rh-menu-item>
-    <rh-menu-item disabled>Disabled Action</rh-menu-item>
-    <rh-menu-item disabled href="#">Disabled link</rh-menu-item>
-    <rh-menu-item disabled>Aria-disabled link</rh-menu-item>
-    <hr/>
-    <rh-menu-item>Separated action</rh-menu-item>
-    <rh-menu-item href="#">Separated link</rh-menu-item>
+    <rh-menu>
+      <rh-menu-item>Action</rh-menu-item>
+      <rh-menu-item href="#">Link</rh-menu-item>
+      <rh-menu-item disabled>Disabled Action</rh-menu-item>
+      <rh-menu-item disabled href="#">Disabled link</rh-menu-item>
+      <rh-menu-item disabled>Aria-disabled link</rh-menu-item>
+      <hr/>
+      <rh-menu-item>Separated action</rh-menu-item>
+      <rh-menu-item href="#">Separated link</rh-menu-item>
+    </rh-menu>
   </rh-menu-dropdown>
 </div>
 
 <script type="module">
   import '@rhds/elements/rh-menu-dropdown/rh-menu-dropdown.js';
-  import '@rhds/elements/rh-menu-dropdown/rh-menu-item.js';
+  import '@rhds/elements/rh-menu/rh-menu-item.js';
+  import '@rhds/elements/rh-menu/rh-menu.js';
 
   const dropdown = document.querySelector('rh-menu-dropdown');
   dropdown.addEventListener('select', e => {
